requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0 
fastapi
uvicorn
pandas
sentence-transformers
faiss-cpu
openai
pydantic_settings